mnist_ftle_3-20-2026/requirements.txt
torch
torchvision
numpy
matplotlib
scipy
scikit-learn
pyyaml
tqdm
